backend/manim-service/requirements.txt
manim==0.19.0
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
